packages/embed-pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embed-pipeline"
version = "0.1.0"
description = "Ray-based parallel embedding pipeline — reads S3, embeds, writes to vector store"
requires-python = ">=3.11"
dependencies = [
    "embed-core",
    "ray[data,default]>=2.20",
    "boto3>=1.34",
    "Pillow>=10.3",
    "pydantic-settings>=2.3",
    "tenacity>=8.3",
    "structlog>=24.1",
    "prefect>=3.1",
    "prefect-aws>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "moto[s3]>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/embed_pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
